Deprecated np float

+++ softgym/envs/pass_water.py
@@ -345,11 +345,11 @@ class PassWater1DEnv(FluidEnv):
             return self.get_image(self.camera_width, self.camera_height)
         elif self.observation_mode == 'point_cloud':
             particle_pos = np.array(pyflex.get_positions()).reshape([-1, 4])[:, :3].flatten()
-            pos = np.zeros(shape=self.particle_obs_dim, dtype=np.float)
+            pos = np.zeros(shape=self.particle_obs_dim, dtype=np.float32)
             pos[:len(particle_pos)] = particle_pos
             return pos.flatten()
         elif self.observation_mode == 'key_point':
-            pos = np.empty(0, dtype=np.float)
+            pos = np.empty(0, dtype=np.float32)
             water_state = pyflex.get_positions().reshape([-1, 4])
             water_num = len(water_state)
             in_glass = self.in_glass(water_state, self.glass_states, self.border, self.height)
